make fp16/bf16 compatible

+++ segment_anything/predictor.py
@@ -228,8 +228,8 @@ class SamPredictor:
         # Predict masks
         low_res_masks, iou_predictions = self.model.mask_decoder(
             image_embeddings=self.features,
-            image_pe=self.model.prompt_encoder.get_dense_pe(),
-            sparse_prompt_embeddings=sparse_embeddings,
+            image_pe=self.model.prompt_encoder.get_dense_pe().type(self.features.dtype),
+            sparse_prompt_embeddings=sparse_embeddings.type(self.features.dtype),
             dense_prompt_embeddings=dense_embeddings,
             multimask_output=multimask_output,
         )
